Split out semgrep-core whole rule running to own class
